MFW-1406 (#12)

+++ services/settings/licensehandler.go
@@ -31,9 +31,9 @@ type LicenseInfo struct {
 type License struct {
 	UID         string `json:"UID"`
 	Type        string `json:"type"`
-	End         int    `json:"end"`
-	Start       int    `json:"start"`
-	Seats       int    `json:"seats" default:"-1"`
+	End         int64  `json:"end"`
+	Start       int64  `json:"start"`
+	Seats       int64  `json:"seats" default:"-1"`
 	DisplayName string `json:"displayName"`
 	Key         string `json:"key"`
 	KeyVersion  int    `json:"keyVersion"`
